% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bassriver}
\alias{bassriver}
\title{Bass River planktic foraminiferal calcite d18O.}
\format{A data frame with 62 rows and 2 column variables:
\describe{
  \item{depth}{marine sediment down-core sample depth, in m}
  \item{d18o}{\emph{Morozovella spp.} shell calcite d18O, in ‰ VPDB}
}}
\source{
John, C. M., S. M. Bohaty, J. C. Zachos, A. Sulijs, S. Gibbs, H. 
Brinkhuis, and T. J. Bralower (2008), North American continental margin 
records of the Paleocene-Eocene thermal maximum: Implications for global 
carbon and hydrological cycling, \emph{Paleoceanography}, 23(2), 
doi:10.1029/2007PA001465.
}
\usage{
bassriver
}
\description{
A data set containing marine sediment core samples from the Paleocene-Eocene 
Thermal Maximum (PETM) section with well preservedforaminifera calcite d18O.
}
\keyword{datasets}
