% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelparams.R
\name{get_available_forams}
\alias{get_available_forams}
\title{Parse trace dataframe column names to get available forams.}
\usage{
get_available_forams(d)
}
\description{
Parse trace dataframe column names to get available forams.
}
