% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.prediction}
\alias{plot.prediction}
\title{Plot a \code{prediction} object.}
\usage{
\method{plot}{prediction}(...)
}
\arguments{
\item{...}{Arguments passed on to \code{predictplot}.}
}
\description{
Plot a \code{prediction} object.
}
\seealso{
\code{\link{predictplot}}
}
