% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_d18oc}
\alias{predict_d18oc}
\title{Predict δ18O of foram calcite given seawater temperature and seawater δ18O.}
\usage{
predict_d18oc(seatemp, d18osw, foram = NULL, seasonal_seatemp = FALSE,
  drawsfun = get_draws)
}
\arguments{
\item{seatemp}{Numeric or vector of observed sea-surface temperatures (°C).}

\item{d18osw}{Numeric or vector of observed seawater δ18O (‰ VSMOW).}

\item{foram}{Optional. String or \code{NULL}. String indicating the foram 
species/subspecies to infer for hierarchical models. String must be one of 
"G. bulloides", "G. ruber white", "G. ruber pink", "G. sacculifer", 
"N. incompta", or "N. pachyderma sinistral". \code{NULL} indicates that a 
pooled model is desired.}

\item{seasonal_seatemp}{Optional boolean indicating whether to use the seasonal 
sea-surface temperature calibrations. Default is \code{FALSE}, i.e. using 
annual SST calibrations.}

\item{drawsfun}{Optional function used to get get model parameter draws. Must 
take arguments for "foram" and "seasonal_seatemp" and return a list with 
members "alpha", "beta", "tau". This is for debugging and testing.}
}
\value{
A \code{prediction} instance for inferred foraminiferal calcite 
δ18O (‰ VPDB).
}
\description{
Predict δ18O of foram calcite given seawater temperature and seawater δ18O.
}
\details{
Four calibration models are available: an "annual pooled" model, a 
"seasonal pooled" model, an "annual hierarchical" model, and a 
"seasonal hierarchical" model. This function uses magic to determine which 
"pooled annual" model is used. Which is the simplest case with potential use 
for Deep Time reconstructions of nonexant foram species. Giving a valid string 
for \code{foram} will use a hierarchical model, which has foram-specific 
variability in calibration model parameters. Passing \code{TRUE} for 
\code{seasonal_seatemp} will use a model trained on season sea-surface 
temperatures. See reference paper for further details.
}
\seealso{
\code{\link{predict_seatemp}}
}
