% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{prediction}
\alias{prediction}
\title{Constructor for S3 prediction class.}
\usage{
prediction(ensemble)
}
\arguments{
\item{ensemble}{A matrix (m x n) of the prediciton posteriors. Where m is the 
number of values inferred and n is the number of trace draws.}
}
\value{
A \code{prediction} object.
}
\description{
Constructor for S3 prediction class.
}
