% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{quantile.prediction}
\alias{quantile.prediction}
\title{Quantiles for a \code{prediction}.}
\usage{
\method{quantile}{prediction}(x, ...)
}
\arguments{
\item{x}{A \code{prediction} object.}

\item{...}{Arguments to be passed on to \code{quantile}.}
}
\description{
Quantiles for a \code{prediction}.
}
