% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{target_timeseries_pred}
\alias{target_timeseries_pred}
\title{Internal function for `predict_seatemp()`.}
\usage{
target_timeseries_pred(d18osw_now, alpha_now, beta_now, tau_now, proxy_ts,
  prior_mu, prior_inv_cov)
}
\value{
Sample of time time series vector conditional on the other args
}
\description{
Internal function for `predict_seatemp()`.
}
